requirements.txt
Flask==2.2.5
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Jinja2==3.1.2
Werkzeug==2.2.3
WTForms~=3.0.1
numpy~=1.25.1
Pillow~=10.1.0
python-dotenv~=1.0.0